Changed int to defined length types

+++ node1/adc.c
@@ -16,20 +16,20 @@ void adc_init(void) {
 	
 }
 
-uint8_t adc_read(int channel) {
+uint8_t adc_read(uint8_t channel) {
 	volatile char* adc_mem = (char*) ADC_ADDRESS;
 	volatile char read_val; //Init read
 	
 	adc_mem[0] = 0; //Activate ADC by writing to it
 	_delay_us(20);
-	for(int i = 0; i <= channel; i++) { //Channel X is by X consecutive writes
+	for(uint8_t i = 0; i <= channel; i++) { //Channel X is by X consecutive writes
 		read_val = adc_mem[0x0];
 	}
 	return read_val;
 }
 
 uint8_t* adc_read_to_array(uint8_t channel, uint8_t* sample_array, uint8_t size){
-	for (int i = 0; i <= size; i++){
+	for (uint8_t i = 0; i <= size; i++){
 		sample_array[i] = adc_read(channel);
 	}
 	return sample_array;
